Move isLowestEnabledRendition to playlist module
