+++ test/master-playlist-controller.test.js
@@ -19,7 +19,6 @@ import { Hls } from '../src/videojs-contrib-hls';
 /* eslint-enable no-unused-vars */
 import Playlist from '../src/playlist';
 import Config from '../src/config';
-import { isLowestEnabledRendition } from '../src/playlist-loader';
 
 const generateMedia = function(isMaat, isMuxed, hasVideoCodec, hasAudioCodec, isFMP4) {
   const codec = (hasVideoCodec ? 'avc1.deadbeef' : '') +
@@ -1454,7 +1453,7 @@ function(assert) {
               1000,
               'default request timeout');
 
-  assert.ok(!isLowestEnabledRendition(
+  assert.ok(!Playlist.isLowestEnabledRendition(
               this.masterPlaylistController.masterPlaylistLoader_.master,
               this.masterPlaylistController.masterPlaylistLoader_.media()),
             'not on lowest rendition');
@@ -1468,7 +1467,7 @@ function(assert) {
   // Download new segment after media change
   this.standardXHRResponse(this.requests[3]);
 
-  assert.ok(isLowestEnabledRendition(
+  assert.ok(Playlist.isLowestEnabledRendition(
               this.masterPlaylistController.masterPlaylistLoader_.master,
               this.masterPlaylistController.masterPlaylistLoader_.media()),
             'on lowest rendition');
